infra/aws-cdk/pyproject.toml
[project]
name = "infra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aws-cdk-lib==2.171.1",
    "constructs>=10.0.0,<11.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
